Abstract identity package tests into separate files
